#16 Move Partitioning related classes to a dedicated package
