+++ euphoria-core/src/test/java/cz/seznam/euphoria/core/client/operator/RepartitionTest.java
@@ -16,7 +16,7 @@
 package cz.seznam.euphoria.core.client.operator;
 
 import cz.seznam.euphoria.core.client.dataset.Dataset;
-import cz.seznam.euphoria.core.client.dataset.HashPartitioner;
+import cz.seznam.euphoria.core.client.dataset.partitioning.HashPartitioner;
 import cz.seznam.euphoria.core.client.flow.Flow;
 import org.junit.Test;
