+++ euphoria-core/src/test/java/cz/seznam/euphoria/core/client/operator/TopPerKeyKeyTest.java
@@ -17,8 +17,8 @@ package cz.seznam.euphoria.core.client.operator;
 
 import com.google.common.collect.Iterables;
 import cz.seznam.euphoria.core.client.dataset.Dataset;
-import cz.seznam.euphoria.core.client.dataset.HashPartitioner;
-import cz.seznam.euphoria.core.client.dataset.HashPartitioning;
+import cz.seznam.euphoria.core.client.dataset.partitioning.HashPartitioner;
+import cz.seznam.euphoria.core.client.dataset.partitioning.HashPartitioning;
 import cz.seznam.euphoria.core.client.dataset.windowing.Time;
 import cz.seznam.euphoria.core.client.flow.Flow;
 import cz.seznam.euphoria.core.client.util.Triple;
