+++ euphoria-fluent/src/main/java/cz/seznam/euphoria/fluent/Dataset.java
@@ -15,7 +15,7 @@
  */
 package cz.seznam.euphoria.fluent;
 
-import cz.seznam.euphoria.core.client.dataset.Partitioner;
+import cz.seznam.euphoria.core.client.dataset.partitioning.Partitioner;
 import cz.seznam.euphoria.core.client.functional.UnaryFunction;
 import cz.seznam.euphoria.core.client.functional.UnaryFunctor;
 import cz.seznam.euphoria.core.client.io.DataSink;
