+++ euphoria-spark/src/main/java/cz/seznam/euphoria/spark/ReduceByKeyTranslator.java
@@ -16,7 +16,7 @@
 package cz.seznam.euphoria.spark;
 
 import com.google.common.base.Preconditions;
-import cz.seznam.euphoria.core.client.dataset.Partitioning;
+import cz.seznam.euphoria.core.client.dataset.partitioning.Partitioning;
 import cz.seznam.euphoria.core.client.dataset.windowing.MergingWindowing;
 import cz.seznam.euphoria.core.client.dataset.windowing.TimedWindow;
 import cz.seznam.euphoria.core.client.dataset.windowing.Window;
